+++ src/server/tcp/interceptors/common/proxyPortsReplacer.ts
@@ -2,6 +2,25 @@ import { Tunnel } from '../../proxyPool';
 import { serverState } from '../../../http/serverState';
 
 class ProxyPortsReplacer {
+  inPortLine(line: string): string {
+    const regExp = /(\d+)/i;
+
+    return line.replace(regExp, (match, port) => {
+      const redisPort = parseInt(port);
+      return this.port(redisPort).toString();
+    });
+  }
+
+  inIpPortLine(line: string): string {
+    const regExp = /(\b\d+\.\d+\.\d+\.\d+:)(\d+)((@\d+)?\b)/i;
+
+    return line.replace(regExp, (match, start, port, end) => {
+      const redisPort = parseInt(port);
+      const debugPort = this.port(redisPort);
+      return start + debugPort.toString() + end;
+    });
+  }
+
   port(port: number): number {
     for (const { src, dst } of this.tunnels()) {
       if (dst === port) {
@@ -9,22 +28,6 @@ class ProxyPortsReplacer {
       }
     }
     return port;
-  }
-
-  inIpPortLine(line: string): string {
-    const regExp = /(\b\d+\.\d+\.\d+\.\d+:)(\d+)((@\d+)?\b)/i;
-
-    return line.replace(regExp, (whole, start, port, end) => {
-      const redisPort = parseInt(port);
-
-      for (const { src, dst } of this.tunnels()) {
-        if (dst === redisPort) {
-          return start + src.toString() + end;
-        }
-      }
-
-      return whole;
-    });
   }
 
   private tunnels(): Tunnel[] {