+++ src/server/tcp/proxyServer.ts
@@ -19,11 +19,9 @@ export class ProxyServer {
 
   private async onConnection(connection: Socket): Promise<void> {
     const redisClient = new RedisClient({ port: this.redisPort });
-    const sideClient = new RedisClient({ port: this.redisPort });
-    const handler = new TrafficHandler(connection, redisClient, sideClient);
+    const handler = new TrafficHandler(connection, redisClient);
 
     await redisClient.connect();
-    await sideClient.connect();
 
     connection.on('close', () => redisClient.end());
     redisClient.on('close', () => connection.end());
@@ -31,7 +29,7 @@ export class ProxyServer {
       console.log('error on redis connection');
       connection.destroy(err);
     });
-    redisClient.on('data', (chunk) => handler.onResponse(chunk));
+    // redisClient.on('data', (chunk) => handler.onResponse(chunk));
     connection.on('data', (chunk) => handler.onRequest(chunk));
   }
 }