+++ src/server/tcp/trafficHandler.ts
@@ -5,66 +5,52 @@ import { EvalShaRequestInterceptor } from './interceptors/evalShaRequestIntercep
 import { EvalRequestInterceptor } from './interceptors/evalRequestInterceptor';
 import { ClusterInterceptor } from './interceptors/clusterInterceptor';
 import { InterceptorChain } from './interceptors/common/interceptorChain';
-import { proxyPortsReplacer } from './interceptors/common/proxyPortsReplacer';
+import { PassInterceptor } from './interceptors/passInterceptor';
+import { InfoInterceptor } from './interceptors/infoInterceptor';
 
 export class TrafficHandler {
   private readonly interceptors: InterceptorChain;
-  private pipelineBuffer: string[] = [];
-  private pipelineQueue = 0;
 
   constructor(
     public readonly connection: Socket,
     private readonly client: RedisClient,
-    public readonly sideClient: RedisClient,
     private monitorTraffic: boolean = true
   ) {
     this.interceptors = new InterceptorChain([
-      new EvalShaRequestInterceptor(this),
-      new EvalRequestInterceptor(this),
-      new ClusterInterceptor(this),
+      new InfoInterceptor(this.client),
+      new EvalShaRequestInterceptor(),
+      new EvalRequestInterceptor(this.client),
+      new ClusterInterceptor(this.client),
+      new PassInterceptor(this.client),
     ]);
   }
 
   async onRequest(chunk: Buffer) {
     if (this.monitorTraffic) {
-      this.logTrafficChunk(chunk.toString(), 'input');
+      this.logTrafficMessage(chunk.toString(), 'input');
     }
 
     const requests = RESP.decodeFull(chunk.toString()) as string[][];
-    this.pipelineBuffer = [];
-    this.pipelineQueue += requests.length;
+
+    let response = '';
     for (const request of requests) {
-      if (!(await this.interceptors.handle(request))) {
-        this.client.write(Buffer.from(RESP.encodeRequest(request)));
+      const requestResponse = await this.interceptors.handle(request);
+
+      if (requestResponse === null) {
+        throw new Error('unhandled request');
       }
-    }
-  }
 
-  onResponse(response: string) {
+      response += requestResponse;
+    }
+
     if (this.monitorTraffic) {
-      this.logTrafficChunk(response, 'output');
+      this.logTrafficMessage(response, 'output');
     }
 
-    if (response.startsWith('-MOVED ')) {
-      response = proxyPortsReplacer.inIpPortLine(response);
-    }
-
-    if (this.pipelineQueue === 0) {
-      console.log('------> sent raw');
-      this.connection.write(response);
-      return;
-    }
-
-    this.pipelineQueue -= 1;
-    this.pipelineBuffer.push(response);
-    if (this.pipelineQueue === 0) {
-      console.log('------> sent');
-      this.connection.write(this.pipelineBuffer.join(''));
-      this.pipelineBuffer = [];
-    }
+    this.connection.write(response);
   }
 
-  private logTrafficChunk(chunk: string, direction: 'input' | 'output') {
+  private logTrafficMessage(chunk: string, direction: 'input' | 'output') {
     if (direction === 'input') {
       console.debug(`[${new Date().toLocaleString()}] --> incoming message`);
     } else {